// src/types/telegram.ts
// src/types/telegram.ts

export interface TelegramUser {
  id: number;
  is_bot: boolean;
  first_name: string;
  last_name?: string;
  username?: string;
  language_code?: string;
}

export interface TelegramChat {
  id: number;
  type: "private" | "group" | "supergroup" | "channel";
  first_name?: string;
  last_name?: string;
  username?: string;
  title?: string;
}

export interface TelegramMessage {

  message_id: number;

  from: TelegramUser;

  chat: TelegramChat;

  date: number;

  text?: string;
}

export interface CallbackQuery {

  id: string;

  from: TelegramUser;

  message: TelegramMessage;

  data?: string;
}

export interface TelegramUpdate {

  update_id: number;

  message?: TelegramMessage;

  callback_query?: CallbackQuery;
}

export interface InlineKeyboardButton {

  text: string;

  callback_data: string;
}

export interface InlineKeyboardMarkup {

  inline_keyboard: InlineKeyboardButton[][];
}